+++ proxy/tun/icmp/packet_test.go
@@ -1,4 +1,4 @@
-package tun
+package icmp
 
 import (
 	stdnet "net"
@@ -9,7 +9,7 @@ import (
 	"gvisor.dev/gvisor/pkg/tcpip/header"
 )
 
-func TestParseICMPEchoRequest(t *testing.T) {
+func TestParseEchoRequest(t *testing.T) {
 	t.Run("ipv4 echo", func(t *testing.T) {
 		var zero tcpip.Address
 		packet := []byte{
@@ -19,11 +19,11 @@ func TestParseICMPEchoRequest(t *testing.T) {
 			0x56, 0x78,
 			0xaa, 0xbb,
 		}
-		if err := rewriteICMPChecksum(header.IPv4ProtocolNumber, packet, zero, zero); err != nil {
+		if err := RewriteChecksum(header.IPv4ProtocolNumber, packet, zero, zero); err != nil {
 			t.Fatal(err)
 		}
 
-		ident, sequence, ok := parseICMPEchoRequest(header.IPv4ProtocolNumber, packet)
+		ident, sequence, ok := ParseEchoRequest(header.IPv4ProtocolNumber, packet)
 		if !ok {
 			t.Fatal("expected ipv4 echo request to parse")
 		}
@@ -42,11 +42,11 @@ func TestParseICMPEchoRequest(t *testing.T) {
 		}
 		src := tcpip.AddrFromSlice([]byte{0x20, 0x01, 0x0d, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1})
 		dst := tcpip.AddrFromSlice([]byte{0x20, 0x01, 0x0d, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2})
-		if err := rewriteICMPChecksum(header.IPv6ProtocolNumber, packet, src, dst); err != nil {
+		if err := RewriteChecksum(header.IPv6ProtocolNumber, packet, src, dst); err != nil {
 			t.Fatal(err)
 		}
 
-		ident, sequence, ok := parseICMPEchoRequest(header.IPv6ProtocolNumber, packet)
+		ident, sequence, ok := ParseEchoRequest(header.IPv6ProtocolNumber, packet)
 		if !ok {
 			t.Fatal("expected ipv6 echo request to parse")
 		}
@@ -56,52 +56,54 @@ func TestParseICMPEchoRequest(t *testing.T) {
 	})
 }
 
-func TestIsMatchingICMPEchoReply(t *testing.T) {
-	ipv4Reply := []byte{
+func TestMatchEchoReply(t *testing.T) {
+	packet := []byte{
 		byte(header.ICMPv4EchoReply), 0,
 		0, 0,
 		0x12, 0x34,
 		0x56, 0x78,
 	}
-	if err := rewriteICMPChecksum(header.IPv4ProtocolNumber, ipv4Reply, tcpip.Address{}, tcpip.Address{}); err != nil {
+	if err := RewriteChecksum(header.IPv4ProtocolNumber, packet, tcpip.Address{}, tcpip.Address{}); err != nil {
 		t.Fatal(err)
 	}
-	if !isMatchingICMPEchoReply(header.IPv4ProtocolNumber, ipv4Reply, 0x1234, 0x5678) {
-		t.Fatal("expected ipv4 echo reply to match")
+
+	if matchedIdent, ok := MatchEchoReply(header.IPv4ProtocolNumber, packet, 0x1234, 0x5678, 0, false); !ok || matchedIdent != 0x1234 {
+		t.Fatalf("expected original identifier match, got ok=%v ident=%x", ok, matchedIdent)
 	}
-	if isMatchingICMPEchoReply(header.IPv4ProtocolNumber, ipv4Reply, 0x1234, 0x5679) {
-		t.Fatal("unexpected ipv4 echo reply match")
+	if matchedIdent, ok := MatchEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678, 0x1234, true); !ok || matchedIdent != 0x1234 {
+		t.Fatalf("expected alternate identifier match, got ok=%v ident=%x", ok, matchedIdent)
+	}
+	if _, ok := MatchEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678, 0, false); ok {
+		t.Fatal("expected mismatched identifier to be rejected without alternate identifier")
 	}
 }
 
-func TestNormalizeICMPEchoReply(t *testing.T) {
-	t.Run("ipv4 raw socket payload with ip header", func(t *testing.T) {
-		reply := []byte{
-			0x45, 0x00, 0x00, 0x1c,
-			0x00, 0x00, 0x00, 0x00,
-			0x40, 0x01, 0x00, 0x00,
-			127, 0, 0, 1,
-			8, 8, 8, 8,
-			byte(header.ICMPv4EchoReply), 0,
-			0, 0,
-			0x12, 0x34,
-			0x56, 0x78,
-		}
+func TestNormalizeEchoReply(t *testing.T) {
+	reply := []byte{
+		0x45, 0x00, 0x00, 0x1c,
+		0x00, 0x00, 0x00, 0x00,
+		0x40, 0x01, 0x00, 0x00,
+		127, 0, 0, 1,
+		8, 8, 8, 8,
+		byte(header.ICMPv4EchoReply), 0,
+		0, 0,
+		0x12, 0x34,
+		0x56, 0x78,
+	}
 
-		normalized, err := normalizeICMPEchoReply(header.IPv4ProtocolNumber, reply)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(normalized) != header.ICMPv4MinimumSize {
-			t.Fatalf("unexpected normalized length: %d", len(normalized))
-		}
-		if !isMatchingICMPEchoReply(header.IPv4ProtocolNumber, normalized, 0x1234, 0x5678) {
-			t.Fatal("expected normalized ipv4 echo reply to match")
-		}
-	})
+	normalized, err := NormalizeEchoReply(header.IPv4ProtocolNumber, reply)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(normalized) != header.ICMPv4MinimumSize {
+		t.Fatalf("unexpected normalized length: %d", len(normalized))
+	}
+	if matchedIdent, ok := MatchEchoReply(header.IPv4ProtocolNumber, normalized, 0x1234, 0x5678, 0, false); !ok || matchedIdent != 0x1234 {
+		t.Fatal("expected normalized ipv4 echo reply to match")
+	}
 }
 
-func TestRewriteICMPChecksum(t *testing.T) {
+func TestRewriteChecksum(t *testing.T) {
 	t.Run("ipv4", func(t *testing.T) {
 		var zero tcpip.Address
 		packet := []byte{
@@ -111,7 +113,7 @@ func TestRewriteICMPChecksum(t *testing.T) {
 			0x56, 0x78,
 			0xaa, 0xbb, 0xcc,
 		}
-		if err := rewriteICMPChecksum(header.IPv4ProtocolNumber, packet, zero, zero); err != nil {
+		if err := RewriteChecksum(header.IPv4ProtocolNumber, packet, zero, zero); err != nil {
 			t.Fatal(err)
 		}
 
@@ -131,7 +133,7 @@ func TestRewriteICMPChecksum(t *testing.T) {
 		}
 		src := tcpip.AddrFromSlice([]byte{0x20, 0x01, 0x0d, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1})
 		dst := tcpip.AddrFromSlice([]byte{0x20, 0x01, 0x0d, 0xb8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2})
-		if err := rewriteICMPChecksum(header.IPv6ProtocolNumber, packet, src, dst); err != nil {
+		if err := RewriteChecksum(header.IPv6ProtocolNumber, packet, src, dst); err != nil {
 			t.Fatal(err)
 		}
 
@@ -149,7 +151,7 @@ func TestRewriteICMPChecksum(t *testing.T) {
 	})
 }
 
-func TestICMPReplyAddrMatches(t *testing.T) {
+func TestReplyAddrMatches(t *testing.T) {
 	tests := []struct {
 		name     string
 		addr     stdnet.Addr
@@ -184,15 +186,15 @@ func TestICMPReplyAddrMatches(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := icmpReplyAddrMatches(tt.addr, tt.expected); got != tt.match {
-				t.Fatalf("icmpReplyAddrMatches() = %v, want %v", got, tt.match)
+			if got := ReplyAddrMatches(tt.addr, tt.expected); got != tt.match {
+				t.Fatalf("ReplyAddrMatches() = %v, want %v", got, tt.match)
 			}
 		})
 	}
 }
 
-func TestDatagramICMPEchoIdentifier(t *testing.T) {
-	ident, ok := datagramICMPEchoIdentifier(&stdnet.UDPAddr{IP: stdnet.IPv4zero, Port: 1234})
+func TestDatagramEchoIdentifier(t *testing.T) {
+	ident, ok := DatagramEchoIdentifier(&stdnet.UDPAddr{IP: stdnet.IPv4zero, Port: 1234})
 	if !ok {
 		t.Fatal("expected UDP addr to produce a datagram icmp identifier")
 	}
@@ -200,44 +202,22 @@ func TestDatagramICMPEchoIdentifier(t *testing.T) {
 		t.Fatalf("unexpected identifier: %d", ident)
 	}
 
-	if _, ok := datagramICMPEchoIdentifier(&stdnet.IPAddr{IP: stdnet.IPv4zero}); ok {
+	if _, ok := DatagramEchoIdentifier(&stdnet.IPAddr{IP: stdnet.IPv4zero}); ok {
 		t.Fatal("expected non-UDP addr to be rejected")
 	}
 }
 
-func TestMatchICMPEchoReply(t *testing.T) {
+func TestRewriteEchoIdentifier(t *testing.T) {
 	packet := []byte{
 		byte(header.ICMPv4EchoReply), 0,
 		0, 0,
 		0x12, 0x34,
 		0x56, 0x78,
 	}
-	if err := rewriteICMPChecksum(header.IPv4ProtocolNumber, packet, tcpip.Address{}, tcpip.Address{}); err != nil {
+	if err := RewriteEchoIdentifier(header.IPv4ProtocolNumber, packet, 0xabcd); err != nil {
 		t.Fatal(err)
 	}
-
-	if matchedIdent, ok := matchICMPEchoReply(header.IPv4ProtocolNumber, packet, 0x1234, 0x5678, 0, false); !ok || matchedIdent != 0x1234 {
-		t.Fatalf("expected original identifier match, got ok=%v ident=%x", ok, matchedIdent)
-	}
-	if matchedIdent, ok := matchICMPEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678, 0x1234, true); !ok || matchedIdent != 0x1234 {
-		t.Fatalf("expected alternate identifier match, got ok=%v ident=%x", ok, matchedIdent)
-	}
-	if _, ok := matchICMPEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678, 0, false); ok {
-		t.Fatal("expected mismatched identifier to be rejected without alternate identifier")
-	}
-}
-
-func TestRewriteICMPEchoIdentifier(t *testing.T) {
-	packet := []byte{
-		byte(header.ICMPv4EchoReply), 0,
-		0, 0,
-		0x12, 0x34,
-		0x56, 0x78,
-	}
-	if err := rewriteICMPEchoIdentifier(header.IPv4ProtocolNumber, packet, 0xabcd); err != nil {
-		t.Fatal(err)
-	}
-	if !isMatchingICMPEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678) {
+	if matchedIdent, ok := MatchEchoReply(header.IPv4ProtocolNumber, packet, 0xabcd, 0x5678, 0, false); !ok || matchedIdent != 0xabcd {
 		t.Fatal("expected rewritten identifier to match")
 	}
 }
